reference-app/frontend/requirements.txt
astroid==2.15.0
backoff==2.2.1
certifi==2022.12.7
chardet==3.0.4
click==8.1.3
Deprecated==1.2.13
Flask==2.2.3
googleapis-common-protos==1.59.0
grpcio==1.51.3
gunicorn==20.0.4
idna==2.10
importlib-metadata==6.0.1
isort==4.3.21
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.6.1
opentelemetry-api==1.17.0
opentelemetry-exporter-otlp==1.17.0
opentelemetry-exporter-otlp-proto-grpc==1.17.0
opentelemetry-exporter-otlp-proto-http==1.17.0
opentelemetry-instrumentation==0.38b0
opentelemetry-instrumentation-flask==0.38b0
opentelemetry-instrumentation-wsgi==0.38b0
opentelemetry-proto==1.17.0
opentelemetry-sdk==1.17.0
opentelemetry-semantic-conventions==0.38b0
opentelemetry-util-http==0.38b0
prometheus-client==0.16.0
prometheus-flask-exporter==0.22.3
protobuf==4.22.1
pycodestyle==2.10.0
pylint==2.4.0
requests==2.23.0
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.25.11
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0
